// old/ProcureHub.WebApp/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        GetPurchaseRequestByIdResponse: {
            /** Format: uuid */
            id: string;
            requestNumber: string;
            title: string;
            description: null | string;
            /** Format: double */
            estimatedAmount: number | string;
            businessJustification: null | string;
            category: components["schemas"]["GetPurchaseRequestByIdCategoryInfo"];
            department: components["schemas"]["GetPurchaseRequestByIdDepartmentInfo"];
            requester: components["schemas"]["GetPurchaseRequestByIdRequesterInfo"];
            status: components["schemas"]["PurchaseRequestStatus"];
            submittedAt: null | components["schemas"]["NullableOfDateTime"];
            reviewedAt: null | components["schemas"]["NullableOfDateTime"];
            reviewedBy: null | components["schemas"]["GetPurchaseRequestByIdReviewerInfo"];
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
// ...
        PurchaseRequestStatus: "Draft" | "Pending" | "Approved" | "Rejected";
        QueryCategoriesResponse: {
            /** Format: uuid */
            id: string;
            name: string;
        };
        QueryDepartmentsResponse: {
            /** Format: uuid */
            id: string;
            name: string;
        };
// ...
        QueryPurchaseRequestsResponse: {
            /** Format: uuid */
            id: string;
            requestNumber: string;
            title: string;
            description: null | string;
            /** Format: double */
            estimatedAmount: number | string;
            businessJustification: null | string;
            category: components["schemas"]["QueryPurchaseRequestsCategoryInfo"];
            department: components["schemas"]["QueryPurchaseRequestsDepartmentInfo"];
            requester: components["schemas"]["QueryPurchaseRequestsRequesterInfo"];
            status: components["schemas"]["PurchaseRequestStatus"];
            submittedAt: null | components["schemas"]["NullableOfDateTime"];
            reviewedAt: null | components["schemas"]["NullableOfDateTime"];
            reviewedBy: null | components["schemas"]["QueryPurchaseRequestsReviewerInfo"];
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}